Lots of changes from SAMBA_2_2: docs/textdocs/PRINTER_DRIVER2.txt 	- add notice to documentation that user must have UID 0 to install a 	  new printer driver
packaging/PHT/TurboLinux/samba2.spec.tmpl
packaging/RedHat/samba2.spec.tmpl
	- update for libnss_wins.so support, prepare for 2.2.0 release

packaging/SuSE/7.1/
packaging/SuSE/7.1/samba-2.2.0-alpha0.dif
packaging/SuSE/7.1/samba.pamd
packaging/SuSE/7.1/samba.spec
	- added SuSE 7.1 packaging

source/lib/interface.c
	- don't memcmp -1 length

source/tdb/tdbutil.c
	- add appropriate cast to return value of va_arg

source/rpc_parse/parse_misc.c
	- ensure that UNISTR's are zero-filled

source/rpc_parse/parse_spoolss.c
	- correction to INFO_6 structure, as it differs between
	  pre-releases of W2K and the shipping build.

source/lib/debug.c
	- add "INFO:" to debug message printed by set debug level message

source/profile/profile.c
source/include/messages.h
source/utils/smbcontrol.c
	- add debugging INFO: messages
	- add message to return profile level when queried

docs/yodldocs/smbcontrol.1.yo
docs/htmldocs/smbcontrol.1.html
docs/manpages/smbcontrol.1
	- added profilelevel to smbcontrol documentation

WHATSNEW.txt
	- added info about profiling code.

source/tests/crypttest.c
	- add crypttest.c so that we can check for truncated crypt on
	  systems that it is relevant for.

source/printing/nt_printing.c
	- canonicalize share names
	- use getpwuid instead of getsmbpwuid

source/rpc_server/srv_spoolss_nt.c
	- canonicalize share names

source/utils/nmblookup.c
	- use boolean variables correctly.

source/smbwrapper/smbw.c
source/smbwrapper/smbw.h
	- added an authentication hook to smbwrapper which allows a (username,
	  workgroup, password) tuple to be provided by another function.

source/lib/messages.c
	- don't send messages to invalid connections.

+++ source/tdb/tdbutil.c
@@ -140,7 +140,7 @@ size_t tdb_pack(char *buf, int bufsize, char *fmt, ...)
 		switch ((c = *fmt++)) {
 		case 'w':
 			len = 2;
-			w = va_arg(ap, uint16);
+			w = (uint16)va_arg(ap, int);
 			if (bufsize >= len) {
 				SSVAL(buf, 0, w);
 			}
